Fixed default play button
Originally, upon first loading the page with no queue or video, the pause button and stop button was displayed. Changed the default buttons to show the play button and the stop button. The fix for this effects the automatic playing for the first video submitted to the queue. The user will have to click play after submitting, but if they will not have to click play for the next song to play once the first song is done. It will still autoplay the next song in queue.